Дорарботка быстрого сохранения и загрузки
Теперь нельзя делать квиксейв во время смерти или паузы. Теперь можно делать квиксейв во время диалога, торговли, лутинга или просто с открытым инвентарём

> Фикс загрузки шагов в MoveNpcTrigger из-за которой Бельгетс не выходил на сцену.

+++ assets/scripts/triggers/MoveNpcTrigger.cs
@@ -205,7 +205,7 @@ public class MoveNpcTrigger: ActivateOtherTrigger
             }
         }
         
-        step = Convert.ToInt16(data["step"]);
+        step = Convert.ToInt32(data["step"]);
         if (step > 0)
         {
             _on_activate_trigger();
